spice-streaming-agent: handle capability negotiation
This patch completes the initial skeleton for capability negotiation.

It populates the outgoing message (CapabilitiesOutMessage) with a bit
array listing, for each known capability, if it is supported or not.

Likewise, it unpacks the capability message received from the server
(InCapabilitiesMessage) and calls the function
`server_capabilities_received` with a vector containing the
capabilities supported by the server.

Signed-off-by: Kevin Pouget <kpouget@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>