ci: Fix error testing under Fedora 31
test-hexdump test uses cmp which is in diffutils package.
In former Fedora versions that package was installed as base
package but on Fedora 31 we need to add it explicitly.
Change the RPM SPEC file in order to fix also package build as
the RPM will call "make check" from the "%check" section.
Also show suite logs if tests failed to make it easier to
fix them in the future.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Snir Sheriber <ssheribe@redhat.com>